Switched doGet to webpack
