+++ package.json
@@ -33,10 +33,8 @@
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-svelte3": "^3.2.1",
     "gulp": "^4.0.2",
-    "gulp-concat": "^2.6.1",
     "gulp-filter": "^7.0.0",
     "gulp-replace": "^1.0.0",
-    "gulp-typescript": "^6.0.0-alpha.1",
     "html-webpack-plugin": "^5.5.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.5.0",
@@ -48,6 +46,7 @@
     "svelte-i18n": "^3.3.13",
     "svelte-loader": "^3.0.0",
     "svelte-preprocess": "^4.9.8",
+    "ts-loader": "^9.2.6",
     "typescript": "^4.5.2",
     "webpack": "^5.64.4",
     "webpack-stream": "^7.0.0"
@@ -56,7 +55,7 @@
     "build": "gulp build",
     "lint:svelte:svelte-check": "svelte-check",
     "lint:svelte": "run-s -c lint:svelte:*",
-    "lint:ts:eslint": "eslint \"src/**/*.svelte\" \"src/**/*.ts\" gulpfile.js frontend.webpack.config.js .eslintrc.js",
+    "lint:ts:eslint": "eslint \"src/**/*.svelte\" \"src/**/*.ts\" gulpfile.js backend.webpack.config.js frontend.webpack.config.js .eslintrc.js",
     "lint:ts": "run-s -c lint:ts:*",
     "lint": "run-s -c lint:*"
   }
