+++ src/backend/d.ts/backend.d.ts
@@ -1,0 +1,1 @@
+declare const global: PublicEndpoints;
