+++ src/backend/doGet.ts
@@ -1,6 +1,4 @@
-/* exported doGet */
-
-function doGet(): GoogleAppsScript.HTML.HtmlOutput {
+export function doGet(): GoogleAppsScript.HTML.HtmlOutput {
   return HtmlService.createTemplateFromFile("index")
     .evaluate()
     .setTitle("Shared Drive Mover");
