+++ src/backend/index.ts
@@ -1,0 +1,3 @@
+import { doGet } from "./doGet";
+
+global.doGet = doGet;
