--- src/frontend/d.ts/gas-functions.d.ts
@@ -1,16 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-declare namespace google.script {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  interface PublicEndpoints {
-    doGet(): GoogleAppsScript.HTML.HtmlOutput;
-    listFolders(parentID: string): Array<NamedRecord>;
-    listSharedDrives(): Array<NamedRecord>;
-    move(
-      folder: string,
-      sharedDrive: string,
-      copyComments: boolean,
-      mergeFolders: boolean,
-      notEmptyOverride: boolean
-    ): void;
-  }
-}
